refactor: removed schemas
