--- src/schemas/index.ts
@@ -1,1 +1,0 @@
-export type { TitleType, RelationRelevance } from './title';
